feat: place sql lambda handler explicitly in a private isolated network
BREAKING CHANGE: previously the default VPC strategy was used, but as
a database should always be in an isolated subnet, this default makes
a lot more sense. But will break code if your VPC does not have an
isolated subnet, and for example only has egress only subnets.

So provider now allows you to specify a vpcSubnet selection.